+++ extend/waibuMpa/theme/component/widget/form-check-toggle.js
@@ -9,7 +9,7 @@ export async function build (handler, params = {}) {
   if (!this.params.attr.label && this.params.attr.name) this.params.attr.label = this.component.req.t(`field.${this.params.attr.name}`)
   const attr = groupAttrs(this.params.attr, ['label', 'hint', 'wrapper'])
   const contents = []
-  attr._.id = this.params.attr.id ?? generateId()
+  attr._.id = this.params.attr.id ?? generateId('alpha')
   if (attr._.color) attr.label.class.push(parseVariant.call(this, { cls: 'btn', value: attr._.color, values: colors, variants, prepend: true }))
   delete attr._.color