+++ extend/waibuMpa/theme/component/widget/scrollspy.js
@@ -10,7 +10,7 @@ async function scrollspy () {
     build = async () => {
       const { isString } = this.app.lib._
       const { generateId } = this.app.lib.aneka
-      const target = isString(this.params.attr.target) ? this.params.attr.target : generateId()
+      const target = isString(this.params.attr.target) ? this.params.attr.target : generateId('alpha')
       this.component.normalizeAttr(this.params, {
         tag: 'div',
         dataBsSpy: 'scroll',